+++ internal/dataplane/kongstate/credentials_test.go
@@ -17,19 +17,23 @@ func TestKeyAuth_SanitizedCopy(t *testing.T) {
 	}{
 		{
 			name: "fills all fields but Consumer and sanitizes key",
-			in: KeyAuth{kong.KeyAuth{
-				Consumer:  &kong.Consumer{Username: kong.String("foo")},
-				CreatedAt: kong.Int(1),
-				ID:        kong.String("2"),
-				Key:       kong.String("3"),
-				Tags:      []*string{kong.String("4.1"), kong.String("4.2")},
-			}},
-			want: KeyAuth{kong.KeyAuth{
-				CreatedAt: kong.Int(1),
-				ID:        kong.String("2"),
-				Key:       kong.String("{vault://52fdfc07-2182-454f-963f-5f0f9a621d72}"),
-				Tags:      []*string{kong.String("4.1"), kong.String("4.2")},
-			}},
+			in: KeyAuth{
+				KeyAuth: kong.KeyAuth{
+					Consumer:  &kong.Consumer{Username: kong.String("foo")},
+					CreatedAt: kong.Int(1),
+					ID:        kong.String("2"),
+					Key:       kong.String("3"),
+					Tags:      []*string{kong.String("4.1"), kong.String("4.2")},
+				},
+			},
+			want: KeyAuth{
+				KeyAuth: kong.KeyAuth{
+					CreatedAt: kong.Int(1),
+					ID:        kong.String("2"),
+					Key:       kong.String("{vault://52fdfc07-2182-454f-963f-5f0f9a621d72}"),
+					Tags:      []*string{kong.String("4.1"), kong.String("4.2")},
+				},
+			},
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
@@ -47,21 +51,25 @@ func TestHMACAuth_SanitizedCopy(t *testing.T) {
 	}{
 		{
 			name: "fills all fields but Consumer and sanitizes secret",
-			in: HMACAuth{kong.HMACAuth{
-				Consumer:  &kong.Consumer{Username: kong.String("foo")},
-				CreatedAt: kong.Int(1),
-				ID:        kong.String("2"),
-				Username:  kong.String("3"),
-				Secret:    kong.String("4"),
-				Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
-			}},
-			want: HMACAuth{kong.HMACAuth{
-				CreatedAt: kong.Int(1),
-				ID:        kong.String("2"),
-				Username:  kong.String("3"),
-				Secret:    redactedString,
-				Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
-			}},
+			in: HMACAuth{
+				HMACAuth: kong.HMACAuth{
+					Consumer:  &kong.Consumer{Username: kong.String("foo")},
+					CreatedAt: kong.Int(1),
+					ID:        kong.String("2"),
+					Username:  kong.String("3"),
+					Secret:    kong.String("4"),
+					Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
+				},
+			},
+			want: HMACAuth{
+				HMACAuth: kong.HMACAuth{
+					CreatedAt: kong.Int(1),
+					ID:        kong.String("2"),
+					Username:  kong.String("3"),
+					Secret:    redactedString,
+					Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
+				},
+			},
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
@@ -79,25 +87,29 @@ func TestJWTAuth_SanitizedCopy(t *testing.T) {
 	}{
 		{
 			name: "fills all fields but Consumer and sanitizes secret",
-			in: JWTAuth{kong.JWTAuth{
-				Consumer:     &kong.Consumer{Username: kong.String("foo")},
-				CreatedAt:    kong.Int(1),
-				ID:           kong.String("2"),
-				Algorithm:    kong.String("3"),
-				Key:          kong.String("4"),
-				RSAPublicKey: kong.String("5"),
-				Secret:       kong.String("6"),
-				Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
-			}},
-			want: JWTAuth{kong.JWTAuth{
-				CreatedAt:    kong.Int(1),
-				ID:           kong.String("2"),
-				Algorithm:    kong.String("3"),
-				Key:          kong.String("4"),
-				RSAPublicKey: kong.String("5"),
-				Secret:       redactedString,
-				Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
-			}},
+			in: JWTAuth{
+				JWTAuth: kong.JWTAuth{
+					Consumer:     &kong.Consumer{Username: kong.String("foo")},
+					CreatedAt:    kong.Int(1),
+					ID:           kong.String("2"),
+					Algorithm:    kong.String("3"),
+					Key:          kong.String("4"),
+					RSAPublicKey: kong.String("5"),
+					Secret:       kong.String("6"),
+					Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
+				},
+			},
+			want: JWTAuth{
+				JWTAuth: kong.JWTAuth{
+					CreatedAt:    kong.Int(1),
+					ID:           kong.String("2"),
+					Algorithm:    kong.String("3"),
+					Key:          kong.String("4"),
+					RSAPublicKey: kong.String("5"),
+					Secret:       redactedString,
+					Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
+				},
+			},
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
@@ -115,21 +127,25 @@ func TestBasicAuth_SanitizedCopy(t *testing.T) {
 	}{
 		{
 			name: "fills all fields but Consumer and sanitizes password",
-			in: BasicAuth{kong.BasicAuth{
-				Consumer:  &kong.Consumer{Username: kong.String("foo")},
-				CreatedAt: kong.Int(1),
-				ID:        kong.String("2"),
-				Username:  kong.String("3"),
-				Password:  kong.String("4"),
-				Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
-			}},
-			want: BasicAuth{kong.BasicAuth{
-				CreatedAt: kong.Int(1),
-				ID:        kong.String("2"),
-				Username:  kong.String("3"),
-				Password:  redactedString,
-				Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
-			}},
+			in: BasicAuth{
+				BasicAuth: kong.BasicAuth{
+					Consumer:  &kong.Consumer{Username: kong.String("foo")},
+					CreatedAt: kong.Int(1),
+					ID:        kong.String("2"),
+					Username:  kong.String("3"),
+					Password:  kong.String("4"),
+					Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
+				},
+			},
+			want: BasicAuth{
+				BasicAuth: kong.BasicAuth{
+					CreatedAt: kong.Int(1),
+					ID:        kong.String("2"),
+					Username:  kong.String("3"),
+					Password:  redactedString,
+					Tags:      []*string{kong.String("5.1"), kong.String("5.2")},
+				},
+			},
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
@@ -147,25 +163,29 @@ func TestOauth2Credential_SanitizedCopy(t *testing.T) {
 	}{
 		{
 			name: "fills all fields but Consumer and sanitizes client secret",
-			in: Oauth2Credential{kong.Oauth2Credential{
-				Consumer:     &kong.Consumer{Username: kong.String("foo")},
-				CreatedAt:    kong.Int(1),
-				ID:           kong.String("2"),
-				Name:         kong.String("3"),
-				ClientID:     kong.String("4"),
-				ClientSecret: kong.String("5"),
-				RedirectURIs: []*string{kong.String("6.1"), kong.String("6.2")},
-				Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
-			}},
-			want: Oauth2Credential{kong.Oauth2Credential{
-				CreatedAt:    kong.Int(1),
-				ID:           kong.String("2"),
-				Name:         kong.String("3"),
-				ClientID:     kong.String("4"),
-				ClientSecret: redactedString,
-				RedirectURIs: []*string{kong.String("6.1"), kong.String("6.2")},
-				Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
-			}},
+			in: Oauth2Credential{
+				Oauth2Credential: kong.Oauth2Credential{
+					Consumer:     &kong.Consumer{Username: kong.String("foo")},
+					CreatedAt:    kong.Int(1),
+					ID:           kong.String("2"),
+					Name:         kong.String("3"),
+					ClientID:     kong.String("4"),
+					ClientSecret: kong.String("5"),
+					RedirectURIs: []*string{kong.String("6.1"), kong.String("6.2")},
+					Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
+				},
+			},
+			want: Oauth2Credential{
+				Oauth2Credential: kong.Oauth2Credential{
+					CreatedAt:    kong.Int(1),
+					ID:           kong.String("2"),
+					Name:         kong.String("3"),
+					ClientID:     kong.String("4"),
+					ClientSecret: redactedString,
+					RedirectURIs: []*string{kong.String("6.1"), kong.String("6.2")},
+					Tags:         []*string{kong.String("7.1"), kong.String("7.2")},
+				},
+			},
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {